requirements/requirements.txt
Django==1.6.1
MySQL-python==1.2.5
PIL==1.1.7
Pygments==1.6
South==0.8.4
argparse==1.2.1
djangorestframework==2.3.12
wsgiref==0.1.2
yolk==0.4.3
